REFACTOR: refatorando a estrutura da aplicação com camadas de responsabilidade
